refactored reachable to be defined in LInkage behavior
refactored reachable to be defined in LInkage behavior
