+++ src/main/scala/behaviors/Linkage.scala
@@ -36,22 +36,21 @@ trait Linkage {
 
   def link(cells: Seq[Cell], bidi: Boolean = true): Seq[Cell] = linkAll(cells, bidi, link)
 
-  // def unlink(cells: Seq[Cell], bidi: Boolean = true): Seq[Cell] = linkAll(cells, bidi, unlink)
-
-  // def link(cells: Seq[Cell], bidi: Boolean = true): Seq[Cell] = {
-  //   linkAll(cells, bidi, link)
-    // val linkedUngrouped: Seq[Cell] = (for ((c1, c2) <- cells zip cells.drop(1)) yield link(c1, c2, bidi)).flatten
-    // val grouped = linkedUngrouped.groupBy(c => (c.coords, c.visited, c.neighbors))
-    // val merged: Seq[Option[Cell]] = grouped.foldLeft(Nil: Seq[Option[Cell]]) {
-    //   case (acc, (k, v)) => {
-    //     val coords: Coordinates = k._1
-    //     val visited: Boolean = k._2
-    //     val neighbors: Neighbors = k._3
-    //     val linked: Set[Coordinates] = v.map(c => c.linked).toSet.flatten
-    //     acc ++ Seq(Some(Cell(coords = coords, visited = visited, neighbors = neighbors, linked = linked)))
-    //   }
-    // }
-    // merged.filter(_.isDefined).map(_.get)
-  // }
+  def reachable(grid: Grid, startX: Int, startY: Int, goalX: Int, goalY: Int): Boolean = {
+    def linked(grid: Grid, coords: Coordinates): Seq[Coordinates] = {
+      var remaining: Seq[Cell] = grid.cells.flatMap(c => c).toSeq
+      var linkedCoords: Seq[Coordinates] = Seq(coords)
+      var currentBatch: Seq[Coordinates] = remaining.filter(c => !c.linked.toSeq.intersect(linkedCoords).isEmpty).map(_.coords)
+      while (currentBatch != Nil) {
+        linkedCoords = linkedCoords ++ currentBatch
+        remaining = remaining.filter(c => !linkedCoords.contains(c.coords))
+        currentBatch = remaining.filter(c => !c.linked.toSeq.intersect(linkedCoords).isEmpty).map(_.coords)
+      }
+      linkedCoords.distinct
+    }
+    val startLinkedCoordinates = linked(grid, Coordinates(startX, startY))
+    val endLinkedCoordinates = linked(grid, Coordinates(goalX, goalY))
+    !startLinkedCoordinates.intersect(endLinkedCoordinates).isEmpty
+  }
 
 }
