+++ src/main/scala/classes/Grid.scala
@@ -51,40 +51,6 @@ case class Grid(
     c.toString * left + s + c.toString * right
   }
 
-  def reachable(startX: Int, startY: Int, goalX: Int, goalY: Int): Boolean = {
-    def linked(grid: Grid, coords: Coordinates): Seq[Coordinates] = {
-      var remaining: Seq[Cell] = grid.cells.flatMap(c => c).toSeq
-      var linkedCoords: Seq[Coordinates] = Seq(coords)
-      var currentBatch: Seq[Coordinates] = remaining.filter(c => !c.linked.toSeq.intersect(linkedCoords).isEmpty).map(_.coords)
-      while (currentBatch != Nil) {
-        linkedCoords = linkedCoords ++ currentBatch
-        remaining = remaining.filter(c => !linkedCoords.contains(c.coords))
-        currentBatch = remaining.filter(c => !c.linked.toSeq.intersect(linkedCoords).isEmpty).map(_.coords)
-      }
-      linkedCoords.distinct
-    }
-    // def linkedOriginal(grid: Grid, coords: Coordinates): Seq[Coordinates] = {
-    //   var linkedCoords: Seq[Coordinates] = Nil
-    //   var frontier: Seq[Coordinates] = Seq(coords)
-    //   while (!frontier.isEmpty) {
-    //     var newFrontier: Seq[Coordinates] = Nil
-    //     for (c <- frontier.map(c => grid.get(c.x)(c.y))) {
-    //       for (linked <- c.linked) {
-    //         if (!linkedCoords.contains(linked)) {
-    //           linkedCoords = linkedCoords ++ Seq(linked)
-    //           newFrontier = newFrontier ++ Seq(linked)
-    //         }
-    //       }
-    //     }
-    //     frontier = newFrontier
-    //   }
-    //   linkedCoords.distinct
-    // }
-    val startLinkedCoordinates = linked(this, Coordinates(startX, startY))
-    val endLinkedCoordinates = linked(this, Coordinates(goalX, goalY))
-    !startLinkedCoordinates.intersect(endLinkedCoordinates).isEmpty
-  }
-
   // given list of Cells, converts list to grid (array of arrays of cells)
   // prerequisite: provided list's length equals our grid's rows multiplied by columns
   def unflatten(flattened: Seq[Cell]): Grid = {
@@ -114,50 +80,6 @@ case class Grid(
   }
   def randomInt(upperBoundary: Int): (Int, RNG) = seed.boundedPositiveInt(upperBoundary)
   def randomBoolean(): (Boolean, RNG) = seed.nextBoolean
-
-  // def distances(cell: Cell): Map[Coordinates, Int] = {
-  //   var distances: Map[Coordinates, Int] = Map(cell.coords -> 0)
-  //   var frontier: Seq[Cell] = Seq(cell)
-  //   while (!frontier.isEmpty) {
-  //     var newFrontier: Seq[Cell] = Nil
-  //     for (c <- frontier) {
-  //       for (linked <- c.linked) {
-  //         if (!distances.keySet.contains(linked)) {
-  //           distances = distances + (linked -> (distances.get(c.coords).getOrElse(0) + 1))
-  //           newFrontier = newFrontier ++ Seq(this.get(linked))
-  //         }
-  //       }
-  //     }
-  //     frontier = newFrontier
-  //   }
-  //   distances
-  // }
-  // def showDistances(x: Int, y: Int): Grid = {
-  //   val dist: Map[Coordinates, Int] = distances(x, y)
-  //   val withDinstances: Seq[Cell] = cells.flatten.map(c => c.copy(value = pad(dist.get(c.coords).getOrElse(" ").toString(), ' ', 3)))
-  //   unflatten(withDinstances)
-  // }
-  // def showDistances(coords: Coordinates): Grid = showDistances(coords.x, coords.y)
-
-  // // def distances(cell: Cell): Map[Coordinates, Int] = {
-  // //   var distances: Map[Coordinates, Int] = Map(cell.coords -> 0)
-  // //   var frontier: Seq[Cell] = Seq(cell)
-  // //   while (!frontier.isEmpty) {
-  // //     var newFrontier: Seq[Cell] = Nil
-  // //     for (c <- frontier) {
-  // //       for (linked <- c.linked) {
-  // //         if (distances.keySet.contains(linked)) {
-  // //           distances = distances + (linked -> (distances.get(c.coords).getOrElse(0) + 1))
-  // //           newFrontier = newFrontier ++ Seq(this.get(linked))
-  // //         }
-  // //       }
-  // //     }
-  // //     frontier = newFrontier
-  // //   }
-  // //   distances
-  // // }
-  // def distances(x: Int, y: Int): Map[Coordinates, Int] = this.distances(this.get(x)(y))
-  // def distances(coords: Coordinates): Map[Coordinates, Int] = this.distances(this.get(coords))
 
   // common monad and other useful functions
   def flatten(): List[Cell] = cells.flatten.toList
